+++ fabric-component-api-v1/build.gradle
