+++ fabric-component-api-v1/src/main/resources/assets/fabric-components-api-v1/icon.png
