+++ fabric-component-api-v1/src/main/resources/fabric-components-api-v1.mixins.json
