+++ fabric-component-api-v1/src/main/resources/fabric.mod.json
