+++ fabric-component-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,18 +1,18 @@
 {
   "schemaVersion": 1,
-  "id": "fabric-components-api-v1-testmod",
+  "id": "fabric-component-api-v1-testmod",
   "name": "Fabric Components API (v1) Test Mod",
   "version": "1.0.0",
   "environment": "*",
   "license": "Apache-2.0",
   "depends": {
     "fabric-data-attachment-api-v1": "*",
-    "fabric-components-api-v1": "*",
+    "fabric-component-api-v1": "*",
     "fabric-lifecycle-events-v1": "*"
   },
   "entrypoints": {
     "main": [
-      "net.fabricmc.fabric.test.components.ComponentTestMod"
+      "net.fabricmc.fabric.test.component.ComponentTestMod"
     ]
   }
 }
