+++ gradle.properties
@@ -21,7 +21,7 @@ fabric-blockrenderlayer-v1-version=1.1.52
 fabric-command-api-v1-version=1.2.49
 fabric-command-api-v2-version=2.2.28
 fabric-commands-v0-version=0.2.66
-fabric-components-api-v1-version=0.1.0
+fabric-component-api-v1-version=0.1.0
 fabric-content-registries-v0-version=8.0.16
 fabric-crash-report-info-v1-version=0.2.29
 fabric-data-attachment-api-v1-version=1.1.28
