+++ settings.gradle
@@ -22,7 +22,7 @@ include 'fabric-block-view-api-v2'
 include 'fabric-blockrenderlayer-v1'
 include 'fabric-client-tags-api-v1'
 include 'fabric-command-api-v2'
-include 'fabric-components-api-v1'
+include 'fabric-component-api-v1'
 include 'fabric-content-registries-v0'
 include 'fabric-convention-tags-v2'
 include 'fabric-crash-report-info-v1'
